Create Form and Initial Data pages to clean HomePage file
